% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn_mcmc.R
\name{bn_mcmc}
\alias{bn_mcmc}
\title{Title}
\usage{
bn_mcmc(X, nodetype, graph, MaxPar = 50L, phi = 1, omega = 6.9,
  InitialNetwork = 2L, drop = 0L, N = 1000L, output = 10L)
}
\arguments{
\item{X}{numeric matrix}

\item{nodetype}{Integer vector. Type of the nodes. 1 = source: 2 = sink;
0 = neither.}

\item{graph}{network object created with \code{create_network()}.}

\item{MaxPar}{Integer. Maximum number of parents allowed for a node.
Default to 50.}

\item{phi}{Numeric. prior on distance from prior network. Defaults to 1.}

\item{omega}{Numeric. prior on network size. Defaults to 6.9.}

\item{InitialNetwork}{Integer. 0 = simulated; 1 = randon; 2 = null.}

\item{drop}{Integer. number to drop for burn-in. Defaults to 0.}

\item{N}{Interger, determines the number of MCMC steps.}

\item{output}{Integer. output every nth iteration. Defaults to 100.}
}
\value{
sss
}
\description{
Title
}
