% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_network}
\alias{fit_network}
\title{Fit bayesian network}
\usage{
fit_network(X, Npar, nodetype, par, Niter, MaxPar = 50L, phi = 1,
  omega = 6.9, InitialNetwork = 2L, drop = 0L, Output = 100L)
}
\arguments{
\item{X}{numeric matrix}

\item{Npar}{Integer vector. Number of parents to each node.}

\item{nodetype}{Integer vector. Type of the nodes. 1 = source: 2 = sink;
0 = neither.}

\item{par}{Integer Matrix. Parents of node p in fitted graph.}

\item{Niter}{Integer. Number of iterations to run in MCMC.}

\item{MaxPar}{Integer. Maximum number of parents allowed for a node.
Default to 50.}

\item{phi}{Numeric. prior on distance from prior network. Defaults to 1.}

\item{omega}{Numeric. prior on network size. Defaults to 6.9.}

\item{InitialNetwork}{Integer. 0 = simulated; 1 = randon; 2 = null.}

\item{drop}{Integer. number to drop for burn-in. Defaults to 0.}

\item{Output}{Integer. output every nth iteration. Defaults to 100.}
}
\description{
Fit bayesian network
}
