% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{main_fun}
\alias{main_fun}
\title{Main network playing function}
\usage{
main_fun(X, Npar, nodetype, par, graph_source, graph_target,
  graph_node_labels, MaxPar = 50L, phi = 1, omega = 6.9,
  InitialNetwork = 2L, drop = 0L, N = 1000L, output = 10L)
}
\arguments{
\item{X}{numeric matrix}

\item{Npar}{Integer vector. Number of parents to each node.}

\item{nodetype}{Integer vector. Type of the nodes. 1 = source: 2 = sink;
0 = neither.}

\item{par}{Integer Matrix. Parents of node p in fitted graph.}

\item{graph_source}{Integer vector. Intergers denoting the source of each
edge.}

\item{graph_target}{Integer vector. Intergers denoting the target of each
edge.}

\item{graph_node_labels}{Integer vector. Denoting the possible values of the
nodes.}

\item{MaxPar}{Integer. Maximum number of parents allowed for a node.
Default to 50.}

\item{phi}{Numeric. prior on distance from prior network. Defaults to 1.}

\item{omega}{Numeric. prior on network size. Defaults to 6.9.}

\item{InitialNetwork}{Integer. 0 = simulated; 1 = randon; 2 = null.}

\item{drop}{Integer. number to drop for burn-in. Defaults to 0.}

\item{N}{Interger, determines the number of MCMC steps.}

\item{output}{Integer. output every nth iteration. Defaults to 100.}
}
\description{
Main network playing function
}
