% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network}
\alias{network}
\title{Synthetic example data}
\format{A list containing a matrix 2000 by 81. And a list with graph
information.}
\usage{
network
}
\description{
A dataset containing 139 out of 157 of H.C. andersens fairy tales
translated in Danish. The UTF-8 plain text was sourced from
http://www.andersenstories.com/ and is divided into elements of up
to about 80 characters each.
}
\keyword{datasets}
