% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{read_dag}
\alias{read_dag}
\title{Read dag file generated from P3 simulation v8}
\usage{
read_dag(file, data, maxpar)
}
\arguments{
\item{file}{file path}

\item{data}{matrix generated by read_data}

\item{maxpar}{maximun number of parents for a node}
}
\value{
names list with 4 elements
}
\description{
temporary helper functions, will not be included in final package.
}
