% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{read_data}
\alias{read_data}
\title{Read data file generated from P3 simulation v8}
\usage{
read_data(file)
}
\arguments{
\item{file}{file path}
}
\value{
matrix
}
\description{
temporary helper functions, will not be included in final package.
}
